moved the tree function to utils and updated references
